[#7136] split entity validation logic
