+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sormastosormas/entities/ReceivedDataProcessor.java
@@ -23,7 +23,9 @@ import javax.ejb.EJB;
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
 
+import de.symeda.sormas.api.i18n.Validations;
 import de.symeda.sormas.api.sormastosormas.immunization.SormasToSormasImmunizationDto;
+import de.symeda.sormas.api.sormastosormas.validation.ValidationErrorMessage;
 import de.symeda.sormas.api.utils.DataHelper;
 import org.apache.commons.collections.CollectionUtils;
 
@@ -38,7 +40,6 @@ import de.symeda.sormas.api.sormastosormas.event.SormasToSormasEventParticipantD
 import de.symeda.sormas.api.sormastosormas.validation.ValidationErrorGroup;
 import de.symeda.sormas.api.sormastosormas.validation.ValidationErrors;
 import de.symeda.sormas.backend.sormastosormas.ValidationHelper;
-import de.symeda.sormas.backend.sormastosormas.data.validation.Sormas2SormasDataValidator;
 import de.symeda.sormas.backend.sormastosormas.entities.caze.ReceivedCaseProcessor;
 import de.symeda.sormas.backend.sormastosormas.entities.contact.ReceivedContactProcessor;
 import de.symeda.sormas.backend.sormastosormas.entities.event.ReceivedEventProcessor;
@@ -51,8 +52,6 @@ import de.symeda.sormas.backend.sormastosormas.share.ShareRequestData;
 @LocalBean
 public class ReceivedDataProcessor {
 
-	@EJB
-	private Sormas2SormasDataValidator dataValidator;
 	@EJB
 	private ReceivedCaseProcessor caseProcessor;
 	@EJB
@@ -74,7 +73,7 @@ public class ReceivedDataProcessor {
 		originInfo.setUuid(DataHelper.createUuid());
 		originInfo.setChangeDate(new Date());
 
-		ValidationErrors originInfoErrors = dataValidator.validateOriginInfo(originInfo, Captions.sormasToSormasOriginInfo);
+		ValidationErrors originInfoErrors = validateOriginInfo(originInfo, Captions.sormasToSormasOriginInfo);
 		if (originInfoErrors.hasError()) {
 			validationErrors.add(originInfoErrors);
 		}
@@ -160,7 +159,7 @@ public class ReceivedDataProcessor {
 		originInfo.setUuid(DataHelper.createUuid());
 		originInfo.setChangeDate(new Date());
 
-		ValidationErrors originInfoErrors = dataValidator.validateOriginInfo(originInfo, Captions.sormasToSormasOriginInfo);
+		ValidationErrors originInfoErrors = validateOriginInfo(originInfo, Captions.sormasToSormasOriginInfo);
 		if (originInfoErrors.hasError()) {
 			validationErrors.add(new ValidationErrors(new ValidationErrorGroup(Captions.sormasToSormasOriginInfo), originInfoErrors));
 		}
@@ -193,6 +192,29 @@ public class ReceivedDataProcessor {
 				validationErrors.add(new ValidationErrors(ValidationHelper.buildEventParticipantValidationGroupName(ep), eventParticipantErrors));
 			}
 		});
+
+		return validationErrors;
+	}
+
+	private ValidationErrors validateOriginInfo(SormasToSormasOriginInfoDto originInfo, String validationGroupCaption) {
+		if (originInfo == null) {
+			return ValidationErrors
+					.create(new ValidationErrorGroup(validationGroupCaption), new ValidationErrorMessage(Validations.sormasToSormasShareInfoMissing));
+		}
+
+		ValidationErrors validationErrors = new ValidationErrors();
+
+		if (originInfo.getOrganizationId() == null) {
+			validationErrors.add(
+					new ValidationErrorGroup(Captions.CaseData_sormasToSormasOriginInfo),
+					new ValidationErrorMessage(Validations.sormasToSormasOrganizationIdMissing));
+		}
+
+		if (DataHelper.isNullOrEmpty(originInfo.getSenderName())) {
+			validationErrors.add(
+					new ValidationErrorGroup(Captions.CaseData_sormasToSormasOriginInfo),
+					new ValidationErrorMessage(Validations.sormasToSormasSenderNameMissing));
+		}
 
 		return validationErrors;
 	}
