+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sormastosormas/entities/event/SormasToSormasEventDtoValidator.java
@@ -1,34 +1,42 @@
 package de.symeda.sormas.backend.sormastosormas.entities.event;
 
+import javax.ejb.LocalBean;
+import javax.ejb.Stateless;
+import javax.inject.Inject;
+
 import de.symeda.sormas.api.event.EventDto;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.sormastosormas.event.SormasToSormasEventDto;
 import de.symeda.sormas.api.sormastosormas.sharerequest.SormasToSormasEventPreview;
 import de.symeda.sormas.api.sormastosormas.validation.ValidationErrors;
-import de.symeda.sormas.backend.sormastosormas.data.validation.Sormas2SormasDataValidator;
+import de.symeda.sormas.backend.sormastosormas.data.infra.InfrastructureValidator;
 import de.symeda.sormas.backend.sormastosormas.data.validation.SormasToSormasDtoValidator;
-
-import javax.ejb.EJB;
-import javax.ejb.LocalBean;
-import javax.ejb.Stateless;
-
 
 @Stateless
 @LocalBean
 public class SormasToSormasEventDtoValidator extends SormasToSormasDtoValidator<EventDto, SormasToSormasEventDto, SormasToSormasEventPreview> {
 
-	@EJB
-	private Sormas2SormasDataValidator dataValidator;
+	public SormasToSormasEventDtoValidator() {
+	}
+
+	@Inject
+	protected SormasToSormasEventDtoValidator(InfrastructureValidator infraValidator) {
+		super(infraValidator);
+	}
 
 	@Override
 	public ValidationErrors validateIncoming(SormasToSormasEventDto sharedData) {
-		return dataValidator.validateEventData(sharedData.getEntity());
+		EventDto event = sharedData.getEntity();
+		ValidationErrors validationErrors = new ValidationErrors();
+		validateLocation(event.getEventLocation(), Captions.CaseData, validationErrors);
+
+		return validationErrors;
 	}
 
 	@Override
 	public ValidationErrors validateIncomingPreview(SormasToSormasEventPreview preview) {
 		ValidationErrors eventValidationErrors = new ValidationErrors();
-		dataValidator.validateLocation(preview.getEventLocation(), Captions.Event, eventValidationErrors);
+		validateLocation(preview.getEventLocation(), Captions.Event, eventValidationErrors);
 		return eventValidationErrors;
 	}
