+++ sormas-backend/src/test/java/de/symeda/sormas/backend/AbstractBeanTest.java
@@ -179,7 +179,6 @@ import de.symeda.sormas.backend.share.ExternalShareInfoFacadeEjb.ExternalShareIn
 import de.symeda.sormas.backend.share.ExternalShareInfoService;
 import de.symeda.sormas.backend.sormastosormas.SormasToSormasFacadeEjb.SormasToSormasFacadeEjbLocal;
 import de.symeda.sormas.backend.sormastosormas.crypto.SormasToSormasEncryptionFacadeEjb;
-import de.symeda.sormas.backend.sormastosormas.data.validation.Sormas2SormasDataValidator;
 import de.symeda.sormas.backend.sormastosormas.entities.caze.SormasToSormasCaseFacadeEjb.SormasToSormasCaseFacadeEjbLocal;
 import de.symeda.sormas.backend.sormastosormas.entities.contact.SormasToSormasContactFacadeEjb.SormasToSormasContactFacadeEjbLocal;
 import de.symeda.sormas.backend.sormastosormas.entities.event.SormasToSormasEventFacadeEjb.SormasToSormasEventFacadeEjbLocal;
@@ -603,10 +602,6 @@ public abstract class AbstractBeanTest extends BaseBeanTest {
 
 	public ShareDataBuilderHelper getShareDataBuilderHelper() {
 		return getBean(ShareDataBuilderHelper.class);
-	}
-
-	public Sormas2SormasDataValidator getSormas2SormasDataValidator() {
-		return getBean(Sormas2SormasDataValidator.class);
 	}
 
 	public ReceivedCaseProcessor getReceivedCaseProcessor() {
